requirements.txt
streamlit==1.32.0
google-cloud-discoveryengine==0.11.2
google-api-core==2.17.1